--- a/doc/release-notes/10814-Differencing improvement.md
+++ b/doc/release-notes/10814-Differencing improvement.md
@@ -1,0 +1,3 @@
+### More Scalable Dataset Version Differencing
+
+Differencing between dataset versions, which is done during dataset edit operations and to populate the dataset page versions table has been made signficantly more scalable.