--- a/src/main/java/edu/harvard/iq/dataverse/DatasetVersionDifference.java
+++ b/src/main/java/edu/harvard/iq/dataverse/DatasetVersionDifference.java
@@ -108,7 +108,7 @@ public final class DatasetVersionDifference {
             }
         }
         long startTime = System.currentTimeMillis();
-        Map<Long, FileMetadata> originalFileMetadataMap = new HashMap<>();
+        Map<Long, FileMetadata> originalFileMetadataMap = new HashMap<>(originalVersion.getFileMetadatas().size());
         Map<Long, FileMetadata> previousIDtoFileMetadataMap = new HashMap<>();
         for (FileMetadata fmdo : originalVersion.getFileMetadatas()) {
             originalFileMetadataMap.put(fmdo.getDataFile().getId(), fmdo);