--- a/src/main/java/edu/harvard/iq/dataverse/settings/FeatureFlags.java
+++ b/src/main/java/edu/harvard/iq/dataverse/settings/FeatureFlags.java
@@ -105,6 +105,10 @@ public enum FeatureFlags {
      * Feature flag for the new Globus upload framework.
      */
     GLOBUS_USE_EXPERIMENTAL_ASYNC_FRAMEWORK("globus-use-experimental-async-framework"),
+    /**
+     * Disable the edit-draft logging of all changes made to a draft dataset
+     */
+    DISABLE_EDIT_DRAFT_LOGGING("disable-edit-draft-logging"),
     ;
     
     final String flag;